+++ test/rubocop/primer_octicon_test.rb
@@ -124,18 +124,42 @@ class RubocopPrimerOcticonTest < CopTest
 
   def test_octicon_with_system_arguments
     investigate(cop, <<-RUBY)
-      octicon(:icon, class: "mr-1")
+      octicon(:icon, class: "mr-1 color-icon-primary")
     RUBY
 
-    assert_correction "primer_octicon(:icon, mr: 1)"
+    assert_correction "primer_octicon(:icon, mr: 1, color: :icon_primary)"
   end
 
   def test_octicon_with_custom_class
     investigate(cop, <<-RUBY)
-      octicon(:icon, class: "mr-1 custom")
+      octicon(:icon, class: "mr-1 color-icon-primary custom")
     RUBY
 
-    assert_correction "primer_octicon(:icon, mr: 1, classes: \"custom\")"
+    assert_correction "primer_octicon(:icon, mr: 1, color: :icon_primary, classes: \"custom\")"
+  end
+
+  def test_converts_text_color_into_icon_color
+    investigate(cop, <<-RUBY)
+      octicon(:icon, class: "mr-1 color-text-primary")
+    RUBY
+
+    assert_correction "primer_octicon(:icon, mr: 1, color: :icon_primary)"
+  end
+
+  def test_converts_text_link_into_icon_info
+    investigate(cop, <<-RUBY)
+      octicon(:icon, class: "mr-1 color-text-link")
+    RUBY
+
+    assert_correction "primer_octicon(:icon, mr: 1, color: :icon_info)"
+  end
+
+  def test_converts_keeps_text_white
+    investigate(cop, <<-RUBY)
+      octicon(:icon, class: "mr-1 color-text-white")
+    RUBY
+
+    assert_correction "primer_octicon(:icon, mr: 1, color: :text_white)"
   end
 
   def test_octicon_with_class_that_cant_be_converted